backend/requirements.txt
# FastAPI & Server
fastapi
uvicorn[standard]
python-multipart

# LangChain & LLM
langchain
langchain-openai
langgraph

# Tavily Search
tavily-python

# Email
resend

# Utilities
pydantic
python-dotenv
httpx
aiohttp
beautifulsoup4

# Database (PostgreSQL)
asyncpg
psycopg2-binary
psycopg[pool]
sqlalchemy[asyncio]
pgvector
langgraph-checkpoint-postgres

# Extraction Engines
firecrawl-py
